포인트관련 책임분리
