+++ src/main/java/kr/hhplus/be/server/service/UserService.java
@@ -19,7 +19,6 @@ public class UserService {
     private final PointHistService pointHistService;
     private final PaymentService paymentService;
 
-    // 사용자조회
     public User getUserInfo(long userId, UserStatus status) {
         return userRepository.findByIdAndStatus(userId, status)
                 .orElseThrow(() -> new CustomException("사용자가 존재하지 않습니다."));
@@ -34,50 +33,5 @@ public class UserService {
 
         return user;
     }
-
-    @Transactional
-    public void deductPointsWithLock(Long userId, long amount) {
-
-        User user = userRepository.findByIdAndStatusWithLock(userId, UserStatus.ACTIVE)
-                .orElseThrow(() -> new CustomException("사용자를 찾을 수 없습니다"));
-
-        user.usePoint(amount);
-    }
-
-    @Transactional
-    public void refundPoints(Long userId, long amount) {
-        User user = userRepository.findByIdAndStatus(userId, UserStatus.ACTIVE)
-                .orElseThrow(() -> new CustomException("사용자를 찾을 수 없습니다: " + userId));
-
-        user.addPoint((int) amount);
-    }
-
-    /**
-     * 포인트 충전
-     */
-    @Transactional
-    public ResponseUserPoint chargePoint(RequestPointCharge requestPointCharge) {
-
-        User user = getUserInfo(requestPointCharge.userId(), UserStatus.ACTIVE);
-
-        User returnUser = paymentService.chargePoint(user, requestPointCharge);
-
-        return ResponseUserPoint.from(returnUser);
-    }
-
-    /**
-     * 포인트 조회
-     */
-    public ResponseUserPoint getPoint(long id) {
-
-        User user = getUserInfo(id, UserStatus.ACTIVE);
-
-        if (user.getPoint() < 0) {
-            throw new RuntimeException("포인트는 음수가 불가능");
-        }
-
-        return ResponseUserPoint.from(user);
-    }
-
 
 }
